refactor(portal): extract application config
This normalizes our setup with that scaffoleded by Angular CLI 16.
